backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3

# Database
psycopg2-binary>=2.9

# Async & Channels
channels>=4.0
channels-redis>=4.2
daphne>=4.0

# Task Queue
celery>=5.3
redis>=5.0

# HTTP Client
httpx>=0.27

# Utilities
python-dotenv>=1.0
networkx>=3.2
